fix: Coverity static analysis issues
- Add missing memory cleanup in decomp_scheme allocation failure path
- Remove obsolete FIXME code blocks in selector modules
- Fix potential stale optarg values in benchmark parameter parsing
- Improve memory allocation error handling with proper cleanup paths
- Add explicit type cast for time() return value

JIRA ticket: CPUPL-7339, CPUPL-7340

+++ test/bench/performance.c
@@ -79,7 +79,7 @@ INT32 run_problem_on_performance_mode(aoclfftz_bench_params_t *params,
     // prepare random seed value
     if (params->use_random_seed)
     {
-        params->seed = time(0);
+        params->seed = (INT32)time(0);
     }
 
 #ifdef AOCL_ENABLE_LOG
